add put strategies

+++ internal/pkg/db/collection/gambler.go
@@ -55,7 +55,7 @@ type BetInfo struct {
 	Response   float64            `json:"response,omitempty"`
 }
 
-func (b *BetInfo) GetResponse() float32 {
+func (b *BetInfo) GetResponse() float64 {
 	// TODO: get from db, make cache
 
 	return 0
